Making sure string returns string for string array
This repairs an issue on the array where when it does perform a JSON.parse() and the items should be strings returning an array of something else that is not an string. This when having a validator it will throw an exception saying val.match is not a function as it can't perform regex on something that is not a string.